Checkpoint work on Guile section
There's still significant work that needs to be done, as evidenced by
the TODO list at the top of the section. Some of the content is
misfiled, some of it needs to be expanded, and some of it needs to be
trimmed. The language between section is inconsistent. There are
probably more problems that need to be addressed.