Give less details in the optargs syntax overview
I was describing the semantics of what #:optional and #:key do in this
section. However, the purpose of this section is to explain the general
syntax of the optargs module. Adding in the semantic explanation is just
noise.

I'm still considering whether or not this section even makes sense. The
only way this saves repetition later on is by avoiding the duplicate
explanation in default values and named values. But there's a lot of
overlap between those two even aside from syntax. I would prefer that
each feature is readable independent of the others, but I'm starting to
think it will be a better text if I just say that default values is
required reading for named values. And the explanation about applying to
all future variables is kind of confusing with #:allow-other-keys and
\#:rest. Maybe I should just have a section that explains how these
features relate to each other, rather than a specific optargs section.