Remove paragraph about default values in named values section
I felt the need to add this paragraph for the sake of completeness,
because the change was technically made to a block of code which is
responsible for handling keyword arguments. However, I am conceiving of
function arguments differently than the authors. In the model presented
by this paper, it might be true that the block of code is generally
concerned with named values but the specific change that was made was in
the logic that handles default values. Mentioning it here adds no value
to this section and potentially increases confusion about the model.