Expand & clarify GOOPS section.
This section feels a little small. I looked through all the commits that
mention GOOPS as well as the dev archive. There's a lot of activity,
it's just not relevant to this paper. The only thing that is relevant is
that the system prints an error if no type signature matches the given
arguments.

I expect to have a lot more to say about GOOPS when I write the
analagous paper about object systems.