Remove TODO notes.
It feels wrong to not present any bytecode in the GOOPS implementation
section. The problem is that the bytecode is just calling the
appropriate functions from the GOOPS module so there's nothing
interesting to say about it. The only thing that's relevant here is that
the signature-searching algorithm will throw an error if no match is
found. The search itself is out of scope.