Rework background section to be more useful.
Remove the explanation of Scheme syntax. The examples are simple enough
that they should be intuitively readable to most programmers, especially
because the example functions are analagous to the ones use in the
sections on other programming languages.

Add explanation of rNrs. There are a couple of places where I refer to
the first report that described something, so this is relevant
background knowledge.

Similarly, add explanation of Guile's evolution in the main background
section instead of describing it ad-hoc in the sections that reference
it.

Completely remove the section about the optargs syntax. This doesn't add
much value because the syntax is explained in the appropriate sections.